Fix Profile picture getting very large
When the "about me" text shifts below the profile picture, as a result of a narrow screen, then the picture blew up to full 75% of the width. This just looked odd, as the picture became very large.
Limiting the max-width to the size used on normal side-by-side view, solves this.